WIP: Major WatchedClasses refactor + add comments
This major refactor removes past classes property, changes how classes
are removed from current, spotsAvailable class property is updated, and how the list of all classes is fetched and used to calculate other properties. This refactor also changes the order in which things are updated so that once the list of allClasses is fetched the completion handler updates the other properties with that data. Marked WIP because this still needs work in order to be able to update the NATVC, currently the vc tries to access it before nowAvailable has been calculated.