Overhaul the push-triggered available class check
didReceiveRemoteNotification will be called when a silent remote push
notification is received. This will then fetch new schedule data and
compare it to previous data to see if new watched classes have become available.
If they are, this will then trigger a local notification that alerts the
user about the new class(es).