Filter out past classes on check for new available
When didReceiveRemoteNotification in AppDelegate checks for newly
available classes, it checks a list of previously known available
classes against the list of those that are newly available to see if
there is a difference. This update means that if the ONLY difference in
those two lists is that the old list contains a class that has past
(which wouldn't show up in the newly available list anyway) that won't
trigger the notification.